fix: Race condition between GamepadPoll detour and ImGui setup
Ideally, we would use
(ImGui.GetIO().ConfigFlags & ImGuiConfigFlags.NavEnableGamepad) > 0
for testing in the GamepadPollDetour whether ImGui should handle gamepad controls
or not. However, this has a race condition during load with the detour which sets
up ImGui and throws if our detour gets called before the other, so we opt-in
for a dedicated boolean.