fix: allow window list edits at any time
Slightly tweak the logic of the WindowSystem Draw() to make a shallow copy of the
window list before iterating over it. This allows for the list to be edited at
any time, but has a trade-off that modifications to the list won't take effect
until the next frame, which is negligible and would have been the case anyways.

This change does not break any existing APIs, and should incur minimal (if any)
performance penalty.