fix: Fix broken case for relative symlinks
Currently, the game will have issues saving files to a relative symlink (for example, when a .DAT file is a symlink pointing to ../CHR_12345/HOTBAR.DAT). This change replaces the symlink resolution system with one that calls `weakly_canonical` to resolve and clean up the path in a single operation.