Add KeyCapture service.
Polling `KeyState` buffer and writing zeros as needed was common method
to implement hotkey feature. This, however, can cause certain actions
like running being abruptly interrupted or misbehaving (processing as if it
was pressed again despite the user never released it) due to the nature
of key state buffer.

This commit introduces `SimulatedKeyState` and `KeyCapture` services.
`SimulatedKeyStates` roughly simulates how the game calculates key states
and `KeyCapture` uses their own rules to decide when to pass this value
as inputs to the game.