Fix multi-line copy/paste between ImGui and XIV
At the moment multi-line copy/paste between ImGui and XIV is broken:

- Going from ImGui -> XIV only copies the first line.
- Going from XIV -> ImGui loses all line breaks.

This happens because XIV uses `\r` line terminators and ImGui uses `\n`.

To fix this we normalize all line breaks that are copied in/out of ImGui
to `\r\n`. This works bidirectionally because ImGui will ignore the `\r`
and XIV will ignore the `\n`.