Fix language selector throwing a exception, use native name for Taiwan
The language selector has only been showing language codes and not the
actual language names since  dd0159ae5a2174819c1541644e5cdbd4ddd98a1d
because "tw" (Taiwan Mandarin) was added and it's not supported by
CultureInfo.

This adds a specific check like the language code to work around this
and stop throwing exceptions. Also converts to a switch so it looks a
bit nicer.