Sanitize IME strings to UCS-2
Our ImGui has ImWChar defined as ushort, and it does not support UCS-2.
ImGui would filter the input characters to UCS-2 range, but our IME
implementation would forcefully set the buffer for the text input. This
filters all candidate and composition strings to fit in UCS-2 range,
notifying the user that some candidates are not properly supported.